feat(eventstream-serde-config-resolver): simplify and add tests
